src/1.Proje/requirements.txt
numpy>=1.24
Pillow>=10.0
PyWavelets>=1.4
scipy>=1.10

